Title Extract spots using information from two channels

Title Extract spots using information from two channels

extract.spots.combined.file(file, folder = "./", thresh.offset = 0.1,
  min.sum.intensity = 2, max.distance = 0.5, use.brightest = FALSE,
  max.spots = 2, full.voxel = FALSE, output = "markers")

Arguments

file

File name

folder

Folder

thresh.offset

Thresh offest used in EBImage::thresh()

min.sum.intensity

spots smaller than min.sum.intensity are ignored

max.distance

use only spots with distance to other color spot smaller than max.distance

use.brightest

Logical; use only brightest in max.distance?

max.spots

maximum of spots (per channel), only when use brightest=TRUE

full.voxel

Logical; output contains full voxel instead of rgb intensities

output

output folder

Value

RGB image with spots will be written to output folder